create supervisor subgraph
