+++ config/config.yml
@@ -1,7 +1,7 @@
 models:
   - name: gpt-4o
-    deployment_id: 123c02e45ca77234
+    deployment_id: d92ed4e1d4bafffc
     temperature: 0
-  - name: gemini-1.0-pro
-    deployment_id: 123d43fd9b87e89b
+  - name: gpt-4o-mini
+    deployment_id: d3da334def2bb9c8
     temperature: 0
